Move ThemeProvider from react-next to react-theme-provider's Compat folder
