+++ packages/react-next/src/Styling.ts
@@ -1,1 +1,2 @@
-export * from './Styling/index';
+export * from 'office-ui-fabric-react/lib/Styling';
+export * from '@fluentui/react-theme-provider/lib/Compat';
