--- packages/react-next/src/Styling/index.ts
@@ -1,3 +1,0 @@
-export * from 'office-ui-fabric-react/lib/Styling';
-
-export * from './ThemeProvider';
