+++ packages/react-theme-provider/package.json
@@ -45,6 +45,7 @@
   "dependencies": {
     "@fluentui/react-stylesheets": "^0.1.13",
     "@uifabric/set-version": "^7.0.18",
+    "@uifabric/styling": "^7.14.5",
     "@uifabric/utilities": "^7.24.5",
     "classnames": "^2.2.6",
     "tslib": "^1.10.0"
