+++ packages/react-theme-provider/src/Compat/ThemeProvider.test.tsx
@@ -1,8 +1,10 @@
 import * as React from 'react';
 import { mount, ReactWrapper } from 'enzyme';
-import { Theme, useTheme, mergeThemes } from '@fluentui/react-theme-provider';
-import { Customizations } from '../Utilities';
-import { FontWeights } from '../Styling';
+import { Theme } from '../types';
+import { useTheme } from '../useTheme';
+import { mergeThemes } from '../mergeThemes';
+import { Customizations } from '@uifabric/utilities';
+import { FontWeights } from '@uifabric/styling';
 import { convertLegacyTheme } from './convertLegacyTheme';
 import { ThemeProvider } from './ThemeProvider';
