+++ packages/react-theme-provider/src/Compat/index.ts
@@ -1,0 +1,3 @@
+export * from './ThemeProvider';
+export * from './useGlobalClassNames';
+export * from './convertLegacyTheme';
