+++ packages/react-theme-provider/src/Compat/useGlobalClassNames.ts
@@ -1,4 +1,5 @@
-import { getGlobalClassNames, GlobalClassNames, useTheme } from '../Styling';
+import { getGlobalClassNames, GlobalClassNames } from '@uifabric/styling';
+import { useTheme } from '../useTheme';
 
 export function useGlobalClassNames<T>(classNames: GlobalClassNames<T>): GlobalClassNames<T> {
   const { tokens: theme } = useTheme();
